read tiff moved to utils
